netninja mern 09 authorization is done

+++ mern_stack_ninja09/frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,21 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 function WorkoutDetails({ workout }) {
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext();
 
     const handleDelete = async () => {
+        if (!user) {
+            return;
+        }
+
         const response = await fetch("/api/workouts/" + workout._id, {
             method: "DELETE",
+            headers: {
+                authorization: `Bearer ${user.token}`,
+            },
         });
 
         const json = await response.json();
